==> Python code/Extra Python Code/28_extracting_hourly_sofa_1056(all_missing_+_non_missing)_septic_nonsepticpatient.py <==
# -*- coding: utf-8 -*-
"""28_Extracting_hourly_sofa_1056(all missing + non missing)_septic_NonSepticPatient.ipynb

Automatically generated by Colaboratory.

Original file is located at
    https://colab.research.google.com/drive/1GkrRzmNm8y4_uk-0QI-vUdkw5zGuhQp_
"""

# Commented out IPython magic to ensure Python compatibility.
!pip install pandas
!pip install request
!pip install psycopg2
import io
import pandas as pd
from IPython.display import display
import matplotlib.pyplot as plt
# %matplotlib inline
import numpy as np
import os
import shutil
import posixpath
#import urllib.request
import datetime
import psycopg2

df_csvdata = pd.read_csv('/Users/safiya.hashmi/MasterThesis/CSV files generated/Only_AllSepsisPatients_with_MissingData_FromSepsisOnset_ToShockOnset_or_SepsisOnset+31h.csv')
# Dataset is now stored in a Pandas Dataframe
print ('shape of original dataframe from CSV : ', df_csvdata.shape)
df_sepsisOnset_SepsisOnsetPlus31h_csvdata = df_csvdata[(df_csvdata['Sepsis_SepsisOnset+ShockOnsetOR31h_timeoverlap_exists']==1) ]

print ('shape of the dataframe after we select only icu stays that have all required signals and timeoverlap from CSV : ', df_sepsisOnset_SepsisOnsetPlus31h_csvdata.shape)

icustay_id_list = df_sepsisOnset_SepsisOnsetPlus31h_csvdata['icustay_id']
print('printing only ICU stay IDs to consider', len( icustay_id_list))

# To connect to the mimic DB
import psycopg2
conn = psycopg2.connect(database="mimic",user="postgres",password="postgres",host="localhost", port="5432")
cur = conn.cursor()
cur.execute("SET search_path TO " + "mimiciii")

import pandas.io.sql as psql
try:
  df_ALL1056_sofadata.drop(df_ALL1056_sofadata.index, inplace=True)
except:
  print('df_ALL1056_sofadata does not exist')
df_ALL1056_sofadata = pd.DataFrame(columns=['SUBJECT_ID','ICUSTAY_ID','SUSPECT_INFECTION_TIME','CALCULATION_TIME','SOFA_SCORE','RESP_SOFA','LIVER_SOFA','RENAL_SOFA','CARDIO_SOFA','CNS_SOFA','COAG_SOFA'])
for icustay_id in icustay_id_list:
  try:
    try:
      df_temp_result.drop(df_temp_result.index, inplace =True)
    except:
      print('df_temp_result does not exist')

    query_getTempData =  """ select distinct
       coh.subject_id as subject_id,
       coh.icustay_id as icustay_id ,
       coh.suspected_infection_time_poe as suspection_of_infection_time,
       a.calculation_time,
       a.sofa_Score,
       a.respiration,
       a.liver,
       a.renal,
       a.cardiovascular,
       a.cns,
       a.coagulation

from sepsis3_cohort coh,
  (
select  bf.icustay_id as icustay_id,
       (bf.calculation_time :: timestamp::date::text ||' '|| extract(hour from bf.calculation_time )::text||':00'||':00')::timestamp as calculation_time,

        bf.sofa_score as sofa_Score,

        bf.respiration as respiration,
        bf.liver as liver,
        bf.renal  as renal,
        bf.cardiovascular as cardiovascular,
        bf.cns as cns,
        bf.coagulation as coagulation

from sepsis3_hourlysofa bf
union
select  en.icustay_id as icustay_id ,
        (en.calculation_time :: timestamp::date::text ||' '|| extract(hour from en.calculation_time )::text||':00'||':00')::timestamp  as calculation_time,

        en.sofa_score as sofa_Score,

        en.respiration as respiration,
        en.liver as liver,
        en.renal  as renal,
        en.cardiovascular as cardiovascular,
        en.cns as cns,
        en.coagulation as coagulation

from sepsis3_hourlysofa_entire_icustay en) a
where a.icustay_id =coh.icustay_id
and coh.icustay_id = """ + str(icustay_id) + "order by a.calculation_time" ; 

    print('before query execution')
    cur.execute(query_getTempData);
    print('query_Executed')
    df_temp_result = pd.DataFrame(cur.fetchall(),columns=['SUBJECT_ID','ICUSTAY_ID','SUSPECT_INFECTION_TIME','CALCULATION_TIME','SOFA_SCORE','RESP_SOFA','LIVER_SOFA','RENAL_SOFA','CARDIO_SOFA','CNS_SOFA','COAG_SOFA'])

    #print(df_temp_result)

    df_ALL1056_sofadata = df_ALL1056_sofadata.append(df_temp_result,ignore_index=True)
    print('Appended SOFA data successfully for ICU STAY : ' ,icustay_id )
  
  except:
    print('Error occurred while appending SOFA data for ICU STAY : ' ,icustay_id )

#print(df_ALL1655_tempdata)
#df_ALL1655_tempdata.to_csv

print(df_ALL1056_sofadata.shape)

df_ALL1056_sofadata.to_csv ('df_ALL1056_sofadata.csv', sep=',', index = False, header=True);
